+++ src/steps/resource-manager/compute/constants.ts
@@ -19,6 +19,8 @@ export const steps = {
   VIRTUAL_MACHINE_EXTENSIONS: 'rm-compute-virtual-machine-extensions',
   VIRTUAL_MACHINE_DISK_RELATIONSHIPS:
     'rm-compute-virtual-machine-disk-relationships',
+  VIRTUAL_MACHINE_IMAGE_RELATIONSHIPS:
+    'rm-compute-virtual-machine-image-relationships',
 };
 
 // Graph object
@@ -71,6 +73,18 @@ export const relationships = {
     sourceType: VIRTUAL_MACHINE_ENTITY_TYPE,
     _class: RelationshipClass.USES,
     targetType: entities.VIRTUAL_MACHINE_EXTENSION._type,
+  },
+  VIRTUAL_MACHINE_USES_IMAGE: {
+    _type: 'azure_vm_uses_image',
+    sourceType: VIRTUAL_MACHINE_ENTITY_TYPE,
+    _class: RelationshipClass.USES,
+    targetType: VIRTUAL_MACHINE_IMAGE_ENTITY_TYPE,
+  },
+  VIRTUAL_MACHINE_USES_SHARED_IMAGE: {
+    _type: 'azure_vm_uses_shared_image',
+    sourceType: VIRTUAL_MACHINE_ENTITY_TYPE,
+    _class: RelationshipClass.USES,
+    targetType: entities.SHARED_IMAGE._type,
   },
   VIRTUAL_MACHINE_USES_UNMANAGED_DISK: {
     _type: 'azure_vm_uses_storage_account',