+++ CHANGELOG.md
@@ -26,6 +26,11 @@ and this project adheres to
   | `azure_vm`             | `HAS`      | `azure_shared_image` |
   | `azure_vm`             | `HAS`      | `azure_image`        |
 
+- Changed the following property values on `azure_vm_extension`:
+  - `settings`
+  - `extType`
+  - `publisher`
+
 ## 5.24.0 - 2021-05-11
 
 ### Added