+++ CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to
 
 ### Added
 
+- Added ingestion of Azure Management Groups when the
+  `configureSubscriptionInstances` configuration field is `true`. This ingestion
+  requires users to assign the `Management Group Reader`
+  [role](https://docs.microsoft.com/en-us/azure/role-based-access-control/built-in-roles#management-group-reader)
+  to the service principal used in the given integration configuration to the
+  **Tenant Root Group**
+  [management group](https://docs.microsoft.com/en-us/azure/governance/management-groups/overview#root-management-group-for-each-directory).
+
 - Added support for ingesting the following **new** resources:
 
   | Service           | Resource / Entity        |